seller functions update
most of the seller functions are done, you can add and edit properties.
i still can't access the codd server so i have no way of testing the images thing, but its based on a method i know works from a different project so in theory it should be fine.